Update ServerMembership validations to enforce implicit member-status of server owners
I had trouble preventing a membership from being deleted if the member was the owner. To handle this, server owners will now be considered members implicitly. Currently, this is handled by the API, in the server jbuilder views. The server owner is added to the list of members. In order to prevent duplication, I added a validation that prevents a server owner from "joining" their own server explicitly